[refactor] 사용자 정보 조회 및 수정 기능 리팩토링 진행(CUL-173)
# 리팩토링 이유
# - 자체적으로 추가한 ApiResponse와 NamedWrapper를 사용하지 않아도 될 것 같아서 수정 진행
# - ApiResponse와 NamedWrapper는 차후 사용할 수 있을 것으로 생각하여 유지 (삭제요청시 삭제하겠습니다)
# 리팩토링 내역
# 1. 사용자 정보 조회 반환값 수정 (UserController.getUserProfile(@Auth User user))
# 2. UserProfile dto 생성 : Json response에서 user로 한 번 감싸주기 위함
# 3. 코드 변경에 따른 UserMapper의 수정
# 4. 응답 예시 리스폰스 삭제 진행
